requirements.txt
netmiko>=4.0
pyserial>=3.5
pyyaml>=6.0
jinja2>=3.0
